backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-google-genai
langchain-community
python-dotenv
httpx
pydantic-settings
google-generativeai
langserve[all]
langchain-core
pytz 